requirements.txt
torch
ultralytics
onnxruntime-gpu
opencv-python
numpy
tqdm
rtmlib
matplotlib
